afsdump warning killing
a couple legit bugs here but mostly i used my parentheses shaker.

Reviewed-on: http://gerrit.openafs.org/1648
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/tests/int64.c
@@ -57,7 +57,7 @@ hexify_int64(u_int64 * X, char *buf)
 #else
     if (!buf)
 	buf = mybuf;
-    sprintf(buf, "%08lx%08lx", X->hi, X->lo);
+    sprintf(buf, "%08lx%08lx", (unsigned long)X->hi, (unsigned long)X->lo);
 #endif
 
     return buf;
